+++ hongit-api/src/main/kotlin/com/devfloor/hongit/api/security/web/AuthorizationType.kt
@@ -1,0 +1,9 @@
+package com.devfloor.hongit.api.security.web
+
+enum class AuthorizationType {
+    BASIC,
+    BEARER,
+    ;
+
+    fun toLowerCase(): String = this.name.lowercase()
+}
