+++ hongit-api/src/main/kotlin/com/devfloor/hongit/api/security/web/resolver/impl/DefaultLoginUserArgumentResolver.kt
@@ -1,12 +1,14 @@
-package com.devfloor.hongit.api.security.web
+package com.devfloor.hongit.api.security.web.resolver.impl
 
+import com.devfloor.hongit.api.security.web.resolver.spec.LoginUserArgumentResolver
 import com.devfloor.hongit.core.user.domain.UserRepository
 import org.springframework.core.MethodParameter
 import org.springframework.web.bind.support.WebDataBinderFactory
 import org.springframework.web.context.request.NativeWebRequest
 import org.springframework.web.method.support.ModelAndViewContainer
+import javax.persistence.EntityNotFoundException
 
-class ProdLoginUserArgumentResolver(
+class DefaultLoginUserArgumentResolver(
     private val userRepository: UserRepository,
 ) : LoginUserArgumentResolver {
     override fun resolveArgument(
@@ -15,6 +17,9 @@ class ProdLoginUserArgumentResolver(
         webRequest: NativeWebRequest,
         binderFactory: WebDataBinderFactory?,
     ): Any? {
-        throw UnsupportedOperationException("아직 구현되지 않은 기능입니다.")
+        val users = userRepository.findAll()
+
+        if (users.isEmpty()) throw EntityNotFoundException("사용자가 존재하지 않습니다.")
+        return users.first()
     }
 }
