+++ hongit-api/src/main/kotlin/com/devfloor/hongit/api/security/web/resolver/impl/ProdLoginUserArgumentResolver.kt
@@ -1,13 +1,13 @@
-package com.devfloor.hongit.api.security.web
+package com.devfloor.hongit.api.security.web.resolver.impl
 
+import com.devfloor.hongit.api.security.web.resolver.spec.LoginUserArgumentResolver
 import com.devfloor.hongit.core.user.domain.UserRepository
 import org.springframework.core.MethodParameter
 import org.springframework.web.bind.support.WebDataBinderFactory
 import org.springframework.web.context.request.NativeWebRequest
 import org.springframework.web.method.support.ModelAndViewContainer
-import javax.persistence.EntityNotFoundException
 
-class DefaultLoginUserArgumentResolver(
+class ProdLoginUserArgumentResolver(
     private val userRepository: UserRepository,
 ) : LoginUserArgumentResolver {
     override fun resolveArgument(
@@ -16,9 +16,6 @@ class DefaultLoginUserArgumentResolver(
         webRequest: NativeWebRequest,
         binderFactory: WebDataBinderFactory?,
     ): Any? {
-        val users = userRepository.findAll()
-
-        if (users.isEmpty()) throw EntityNotFoundException("사용자가 존재하지 않습니다.")
-        return users.first()
+        throw UnsupportedOperationException("아직 구현되지 않은 기능입니다.")
     }
 }
