+++ hongit-api/src/test/kotlin/com/devfloor/hongit/api/support/ApiDocsTestUtils.kt
@@ -1,6 +1,6 @@
 package com.devfloor.hongit.api.support
 
-import com.devfloor.hongit.api.security.web.DefaultLoginUserArgumentResolver
+import com.devfloor.hongit.api.security.web.resolver.impl.DefaultLoginUserArgumentResolver
 import com.devfloor.hongit.core.user.domain.UserRepository
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter
